move PieGraph to Components folder
